bug 11518: Return "saved_password" for non-existed password entry in domR
To solve password file is destroyed along with restartNetwork command issue. If
the password is not set in fact, user can use "ResetPassword" to try again. But
it won't happen mostly, because it's only possible if the restartNetwork
happened between user start up VM and set the new password.

Reviewed-by: Keshav

status 11518: resolved fixed